# supabase/schema-openapi.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "title": "standard public schema",
    "version": "14.1"
  },
  "host": "wjcbobcqyqdkludsbqgf.supabase.co:443",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json",
    "application/vnd.pgrst.object+json;nulls=stripped",
    "application/vnd.pgrst.object+json",
    "text/csv"
  ],
  "produces": [
    "application/json",
    "application/vnd.pgrst.object+json;nulls=stripped",
    "application/vnd.pgrst.object+json",
    "text/csv"
  ],
  "paths": {
...
    "/noticias": {
      "get": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.noticias.id"
          },
          {
            "$ref": "#/parameters/rowFilter.noticias.titulo"
          },
          {
            "$ref": "#/parameters/rowFilter.noticias.descricao"
          },
          {
            "$ref": "#/parameters/rowFilter.noticias.conteudo"
          },
          {
            "$ref": "#/parameters/rowFilter.noticias.imagem"
          },
          {
            "$ref": "#/parameters/rowFilter.noticias.autorid"
          },
          {
            "$ref": "#/parameters/rowFilter.noticias.destaque"
          },
          {
            "$ref": "#/parameters/rowFilter.noticias.publica"
          },
          {
            "$ref": "#/parameters/rowFilter.noticias.datapublicacao"
          },
          {
            "$ref": "#/parameters/rowFilter.noticias.datacriacao"
          },
          {
            "$ref": "#/parameters/rowFilter.noticias.dataatualizacao"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/order"
          },
          {
            "$ref": "#/parameters/range"
          },
          {
            "$ref": "#/parameters/rangeUnit"
          },
          {
            "$ref": "#/parameters/offset"
          },
          {
            "$ref": "#/parameters/limit"
          },
          {
            "$ref": "#/parameters/preferCount"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "items": {
                "$ref": "#/definitions/noticias"
              },
              "type": "array"
            }
          },
          "206": {
            "description": "Partial Content"
          }
        },
        "tags": [
          "noticias"
        ]
      },
      "post": {
        "parameters": [
          {
            "$ref": "#/parameters/body.noticias"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/preferPost"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "tags": [
          "noticias"
        ]
      },
      "delete": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.noticias.id"
          },
          {
            "$ref": "#/parameters/rowFilter.noticias.titulo"
          },
          {
            "$ref": "#/parameters/rowFilter.noticias.descricao"
          },
          {
            "$ref": "#/parameters/rowFilter.noticias.conteudo"
          },
          {
            "$ref": "#/parameters/rowFilter.noticias.imagem"
          },
          {
            "$ref": "#/parameters/rowFilter.noticias.autorid"
          },
          {
            "$ref": "#/parameters/rowFilter.noticias.destaque"
          },
          {
            "$ref": "#/parameters/rowFilter.noticias.publica"
          },
          {
            "$ref": "#/parameters/rowFilter.noticias.datapublicacao"
          },
          {
            "$ref": "#/parameters/rowFilter.noticias.datacriacao"
          },
          {
            "$ref": "#/parameters/rowFilter.noticias.dataatualizacao"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "noticias"
        ]
      },
      "patch": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.noticias.id"
          },
          {
            "$ref": "#/parameters/rowFilter.noticias.titulo"
          },
          {
            "$ref": "#/parameters/rowFilter.noticias.descricao"
          },
          {
            "$ref": "#/parameters/rowFilter.noticias.conteudo"
          },
          {
            "$ref": "#/parameters/rowFilter.noticias.imagem"
          },
          {
            "$ref": "#/parameters/rowFilter.noticias.autorid"
          },
          {
            "$ref": "#/parameters/rowFilter.noticias.destaque"
          },
          {
            "$ref": "#/parameters/rowFilter.noticias.publica"
          },
          {
            "$ref": "#/parameters/rowFilter.noticias.datapublicacao"
          },
          {
            "$ref": "#/parameters/rowFilter.noticias.datacriacao"
          },
          {
            "$ref": "#/parameters/rowFilter.noticias.dataatualizacao"
          },
          {
            "$ref": "#/parameters/body.noticias"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "noticias"
        ]
      }
    },
...
    "/funcionarios": {
      "get": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.funcionarios.id"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.usuarioid"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.matricula"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.funcao"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.rg"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.telefonecelular"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.endereco"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.numero"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.bairro"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.cep"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.cidade"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.uf"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.dtnascimento"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.dtadmissao"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.status"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.banco"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.agencia"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.contacorrente"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.pix"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.obs"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.datacriacao"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.dataatualizacao"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.numero_id"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/order"
          },
          {
            "$ref": "#/parameters/range"
          },
          {
            "$ref": "#/parameters/rangeUnit"
          },
          {
            "$ref": "#/parameters/offset"
          },
          {
            "$ref": "#/parameters/limit"
          },
          {
            "$ref": "#/parameters/preferCount"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "items": {
                "$ref": "#/definitions/funcionarios"
              },
              "type": "array"
            }
          },
          "206": {
            "description": "Partial Content"
          }
        },
        "tags": [
          "funcionarios"
        ]
      },
      "post": {
        "parameters": [
          {
            "$ref": "#/parameters/body.funcionarios"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/preferPost"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "tags": [
          "funcionarios"
        ]
      },
      "delete": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.funcionarios.id"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.usuarioid"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.matricula"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.funcao"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.rg"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.telefonecelular"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.endereco"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.numero"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.bairro"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.cep"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.cidade"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.uf"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.dtnascimento"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.dtadmissao"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.status"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.banco"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.agencia"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.contacorrente"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.pix"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.obs"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.datacriacao"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.dataatualizacao"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.numero_id"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "funcionarios"
        ]
      },
      "patch": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.funcionarios.id"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.usuarioid"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.matricula"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.funcao"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.rg"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.telefonecelular"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.endereco"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.numero"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.bairro"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.cep"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.cidade"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.uf"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.dtnascimento"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.dtadmissao"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.status"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.banco"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.agencia"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.contacorrente"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.pix"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.obs"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.datacriacao"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.dataatualizacao"
          },
          {
            "$ref": "#/parameters/rowFilter.funcionarios.numero_id"
          },
          {
            "$ref": "#/parameters/body.funcionarios"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "funcionarios"
        ]
      }
    },
...
  },
  "definitions": {
...
    "noticias": {
      "required": [
        "id",
        "titulo"
      ],
      "properties": {
        "id": {
          "description": "Note:\nThis is a Primary Key.<pk/>",
          "format": "integer",
          "type": "integer"
        },
        "titulo": {
          "format": "character varying",
          "maxLength": 255,
          "type": "string"
        },
        "descricao": {
          "format": "text",
          "type": "string"
        },
        "conteudo": {
          "format": "text",
          "type": "string"
        },
        "imagem": {
          "format": "text",
          "type": "string"
        },
        "autorid": {
          "description": "Note:\nThis is a Foreign Key to `usuarios.id`.<fk table='usuarios' column='id'/>",
          "format": "integer",
          "type": "integer"
        },
        "destaque": {
          "default": false,
          "format": "boolean",
          "type": "boolean"
        },
        "publica": {
          "default": true,
          "format": "boolean",
          "type": "boolean"
        },
        "datapublicacao": {
          "default": "CURRENT_TIMESTAMP",
          "format": "timestamp without time zone",
          "type": "string"
        },
        "datacriacao": {
          "default": "CURRENT_TIMESTAMP",
          "format": "timestamp without time zone",
          "type": "string"
        },
        "dataatualizacao": {
          "default": "CURRENT_TIMESTAMP",
          "format": "timestamp without time zone",
          "type": "string"
        }
      },
      "type": "object"
    },
...
    "funcionarios": {
      "required": [
        "id",
        "funcao",
        "numero_id"
      ],
      "properties": {
        "id": {
          "description": "Note:\nThis is a Primary Key.<pk/>",
          "format": "integer",
          "type": "integer"
        },
        "usuarioid": {
          "description": "Note:\nThis is a Foreign Key to `usuarios.id`.<fk table='usuarios' column='id'/>",
          "format": "integer",
          "type": "integer"
        },
        "matricula": {
          "format": "character varying",
          "maxLength": 50,
          "type": "string"
        },
        "funcao": {
          "format": "character varying",
          "maxLength": 100,
          "type": "string"
        },
        "rg": {
          "format": "character varying",
          "maxLength": 20,
          "type": "string"
        },
        "telefonecelular": {
          "format": "character varying",
          "maxLength": 20,
          "type": "string"
        },
        "endereco": {
          "format": "character varying",
          "maxLength": 255,
          "type": "string"
        },
        "numero": {
          "format": "character varying",
          "maxLength": 10,
          "type": "string"
        },
        "bairro": {
          "format": "character varying",
          "maxLength": 100,
          "type": "string"
        },
        "cep": {
          "format": "character varying",
          "maxLength": 10,
          "type": "string"
        },
        "cidade": {
          "format": "character varying",
          "maxLength": 100,
          "type": "string"
        },
        "uf": {
          "format": "character",
          "maxLength": 2,
          "type": "string"
        },
        "dtnascimento": {
          "format": "date",
          "type": "string"
        },
        "dtadmissao": {
          "format": "date",
          "type": "string"
        },
        "status": {
          "default": "ATIVO",
          "format": "character varying",
          "maxLength": 50,
          "type": "string"
        },
        "banco": {
          "format": "character varying",
          "maxLength": 50,
          "type": "string"
        },
        "agencia": {
          "format": "character varying",
          "maxLength": 10,
          "type": "string"
        },
        "contacorrente": {
          "format": "character varying",
          "maxLength": 20,
          "type": "string"
        },
        "pix": {
          "format": "character varying",
          "maxLength": 100,
          "type": "string"
        },
        "obs": {
          "format": "text",
          "type": "string"
        },
        "datacriacao": {
          "default": "CURRENT_TIMESTAMP",
          "format": "timestamp without time zone",
          "type": "string"
        },
        "dataatualizacao": {
          "default": "CURRENT_TIMESTAMP",
          "format": "timestamp without time zone",
          "type": "string"
        },
        "numero_id": {
          "format": "integer",
          "type": "integer"
        }
      },
      "type": "object"
    },
...
  },
  "parameters": {
...
    "preferReturn": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "return=representation",
        "return=minimal",
        "return=none"
      ],
      "in": "header",
      "type": "string"
    },
    "preferCount": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "count=none"
      ],
      "in": "header",
      "type": "string"
    },
    "preferPost": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "return=representation",
        "return=minimal",
        "return=none",
        "resolution=ignore-duplicates",
        "resolution=merge-duplicates"
      ],
      "in": "header",
      "type": "string"
    },
    "select": {
      "name": "select",
      "description": "Filtering Columns",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
    "order": {
      "name": "order",
      "description": "Ordering",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "range": {
      "name": "Range",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "header",
      "type": "string"
    },
    "rangeUnit": {
      "name": "Range-Unit",
      "description": "Limiting and Pagination",
      "required": false,
      "default": "items",
      "in": "header",
      "type": "string"
    },
    "offset": {
      "name": "offset",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "limit": {
      "name": "limit",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
    "body.noticias": {
      "name": "noticias",
      "description": "noticias",
      "required": false,
      "in": "body",
      "schema": {
        "$ref": "#/definitions/noticias"
      }
    },
    "rowFilter.noticias.id": {
      "name": "id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.noticias.titulo": {
      "name": "titulo",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.noticias.descricao": {
      "name": "descricao",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.noticias.conteudo": {
      "name": "conteudo",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.noticias.imagem": {
      "name": "imagem",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.noticias.autorid": {
      "name": "autorid",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.noticias.destaque": {
      "name": "destaque",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.noticias.publica": {
      "name": "publica",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.noticias.datapublicacao": {
      "name": "datapublicacao",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.noticias.datacriacao": {
      "name": "datacriacao",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.noticias.dataatualizacao": {
      "name": "dataatualizacao",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
    "body.funcionarios": {
      "name": "funcionarios",
      "description": "funcionarios",
      "required": false,
      "in": "body",
      "schema": {
        "$ref": "#/definitions/funcionarios"
      }
    },
    "rowFilter.funcionarios.id": {
      "name": "id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.funcionarios.usuarioid": {
      "name": "usuarioid",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.funcionarios.matricula": {
      "name": "matricula",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.funcionarios.funcao": {
      "name": "funcao",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.funcionarios.rg": {
      "name": "rg",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.funcionarios.telefonecelular": {
      "name": "telefonecelular",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.funcionarios.endereco": {
      "name": "endereco",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.funcionarios.numero": {
      "name": "numero",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.funcionarios.bairro": {
      "name": "bairro",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.funcionarios.cep": {
      "name": "cep",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.funcionarios.cidade": {
      "name": "cidade",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.funcionarios.uf": {
      "name": "uf",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.funcionarios.dtnascimento": {
      "name": "dtnascimento",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.funcionarios.dtadmissao": {
      "name": "dtadmissao",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.funcionarios.status": {
      "name": "status",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.funcionarios.banco": {
      "name": "banco",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.funcionarios.agencia": {
      "name": "agencia",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.funcionarios.contacorrente": {
      "name": "contacorrente",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.funcionarios.pix": {
      "name": "pix",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.funcionarios.obs": {
      "name": "obs",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.funcionarios.datacriacao": {
      "name": "datacriacao",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.funcionarios.dataatualizacao": {
      "name": "dataatualizacao",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.funcionarios.numero_id": {
      "name": "numero_id",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
  },
...
}